+++ pom.xml
@@ -9,8 +9,8 @@
     <version>1.0-SNAPSHOT</version>
 
     <properties>
-        <maven.compiler.target>1.14</maven.compiler.target>
-        <maven.compiler.source>1.14</maven.compiler.source>
+        <maven.compiler.target>1.11</maven.compiler.target>
+        <maven.compiler.source>1.11</maven.compiler.source>
     </properties>
 
     <dependencies>
