+++ src/main/java/element/CartModal.java
@@ -1,0 +1,21 @@
+package element;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.WebDriverWait;
+
+public class CartModal extends AbstractElement {
+
+    private final By header = By.cssSelector("#layer_cart div.layer_cart_product h2");
+
+    public CartModal(WebDriver driver) {
+        super(driver);
+    }
+
+    public String getSuccessHeader() {
+        WebDriverWait wait = new WebDriverWait(driver, TIMEOUT);
+        wait.until(ExpectedConditions.visibilityOfElementLocated(header));
+        return driver.findElement(header).getText();
+    }
+}
