+++ src/main/java/navigation/Navigable.java
@@ -1,0 +1,5 @@
+package navigation;
+
+public interface Navigable {
+    TopNavigation provideNavigation();
+}
