+++ src/test/java/login/LoginTests.java
@@ -2,6 +2,7 @@ package login;
 
 import base.BaseTests;
 import org.testng.annotations.Test;
+import page.AuthenticationPage;
 
 import static org.testng.Assert.assertTrue;
 
@@ -9,9 +10,9 @@ public class LoginTests extends BaseTests {
 
     @Test
     public void testSignIn() {
-        var loginPage = homePage.clickSignIn();
+        AuthenticationPage loginPage = homePage.clickSignIn();
         loginPage.logIn("9jsmith@garcia.com", "6c1Ca");
-        assertTrue(loginPage.isSignedIn(), "No SignOut button: probably not signed in");
+        assertTrue(loginPage.provideNavigation().isSignedIn(), "No SignOut button: probably not signed in");
     }
 
     @Test(dependsOnMethods = "testSignIn")
